refactor: lean on the git binary more
because libgit2 has made it tough to implement support for

- signed commits, as well as
- unsigned commits

- ssh remotes, as well as
- https or local remotes

The git binary is the result of lots of effort in these areas, so
should act as an understandable and portable alternative to git2 fu,
either in the short term or long-term.

Downsides:

- tougher testing?
- more difficuty resuming after an issue?
- stdin cannot be piped to these processes

I'm not sure how much any of these trade-offs matter right now; the
most important thing is to get this working.