feat!: add --separate option
To avoid two options with the short `-s` flag, this commit renames
`--since` to `--base`. This seems to have a more consistent meaning
within git projects.

Add a `--separate` flag that does not group any commits, not even
commits sharing the same issue footer. This can be useful if you want
to split apart one feature into multiple PRs, for example when different
code-owners are involved.