feat: implement overlay
Closes #182

The `overlay` flag combines multiple issue groups into one PR.

This is a strange case for git-disjoint, a tool primarily designed
to separate changes into smaller PRs, to handle.

However, sometimes it is useful given underlying repository constraints.

When `overlay` is activated using the `--overlay` flag, the `choose`
menu prompts the user to select which issue-groups to combine into
a single PR.

When the `overlay` flag is active, git-disjoint is capable of making
only a single PR.